refactor: myProducts refactoring
