+++ my-monorepo/apps/shopease/src/pages/myProducts/hooks/useHandleCategory.ts
@@ -1,0 +1,23 @@
+import { Product } from "@/features/product/api/types";
+import { useCallback } from "react";
+
+export type ProductCategoryType = Product["productCategory"];
+
+type HandleCategoryProps = {
+  setSelectedCategory: React.Dispatch<
+    React.SetStateAction<ProductCategoryType>
+  >;
+};
+
+export const useHandleCategory = ({
+  setSelectedCategory,
+}: HandleCategoryProps) => {
+  const handleCategory = useCallback(
+    (category: ProductCategoryType) => {
+      setSelectedCategory(category);
+    },
+    [setSelectedCategory],
+  );
+
+  return { handleCategory };
+};
